Can render individual character bitmaps

+++ src/gfx/texture.h
@@ -9,7 +9,7 @@ struct Texture {
 	VkDeviceMemory memory;
 };
 
-struct Texture texture_new(uint32* pxs, int w, int h);
+struct Texture texture_new(uint8* pxs, int w, int h);
 struct Texture texture_new_from_image(const char* path);
 void texture_free(struct Texture tex);
